Initial commit.
This robot code is designed for the cRIO control system, and can be build with the Linux cRIO toolchain.
It is intended to be the final robot code for the 2011 robot, Master Chief.
A binary is included for direct deployment, and can be uploaded to the cRIO with FTP if necessary.
Port numbers can be found in Values.h, and should not change. If necessary, the robot should be rewired to match the code.
PID is disabled by default, but can be enabled via the drive joysticks if desired.

Controls:
	Joystick 1 (right):
		X-axis: Strafe
		Y-axis: Forward
		Trigger: Stop
		Button 2: Drive PID On
		Button 3: Drive PID Off
	Joystick 2 (left):
		X-axis: Turn
		Trigger: Stop
		Button 2: Drive PID On
		Button 3: Drive PID Off
	Joystick 3 (lift):
		Y-axis: Elevator
		Trigger: Claw (grab)
		Button 3: Claw (release)
		Button 4: Minibot (extend)
		Button 5: Minibot (retract)